add valued condition abstraction layer
add layer on top of whereClause to be able to retrieve valued vars

This solution introduce some complexity that might affect the
performance of the query builder and retrieving the var.

Therefore we should not consider this solution as a final solution
but rather an example of how to isolate concerns with interfaces with
Go. There must be some work to be done if we want to optimize both
the builder and the getter of the query.